Format Str as 'Str "value"' and Space as 'Space'
Previously, Str was formatted as 'Str text: "value"' but that was
inconsistent because it should really have been 'Str {text: "value"}'.
Given the widespread usage of Str, and the fact that its constructor
can be given a string, it makes sense to hide the 'text' attribute name.
The same logic would also apply to other types with single 'text'
attributes.

Similarly, Space was formatted as 'Space {}' but it's so common that
it makes sense to omit its empty contents.